feat: Update CI/CD workflow trigger action and upgrade the gh-pages version
Now, the workflow will only trigger when changes are pushed to main and not on creation of pull request. Also upgraded the peaceiris/actions-gh-pages to version 4 from version 3. This should suppress the build warning about old node.js version and it will also use the latest version of the repo.